### 2024-02-09 #628 @abigailkeller
eDNAjoint: an R package for interpreting paired environmental DNA and traditional surveys
Submitting Author Name: Abigail Keller
Submitting Author Github Handle: @abigailkeller
Repository:  https://github.com/abigailkeller/eDNAjoint
Submission type: Pre-submission

---

```
Version: 0.1
Maintainer: Abigail G. Keller <agkeller@berkeley.edu>
    c(person("Abigail G.", "Keller", role = c("aut", "cre"), email="agkeller@berkeley.edu"),
    person("Ryan P.", "Kelly", role = "ctb", email="rpkelly@uw.edu"))
Description: Models integrate environmental DNA (eDNA) detection data and traditional survey data to jointly estimate species catch rate. Models can be used with count data via traditional survey methods (i.e., trapping, electrofishing, visual) and replicated eDNA detection/nondetection data via polymerase chain reaction (i.e., PCR or qPCR) from multiple survey locations. Estimated parameters include probability of a false positive eDNA detection, a site-level covariates that scale the sensitivity of eDNA surveys relative to traditional surveys, and catchability coefficients for traditional gear types. Models are implemented with a Bayesian framework (Markov chain Monte Carlo) using the 'Stan' probabilistic programming language.
Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
Biarch: true
    bayestestR,
    loo,
    Rcpp (>= 0.12.0),
    RcppParallel (>= 5.0.1),
    rstan (>= 2.26.23),
    rstantools (>= 2.3.1.1),
    tidyr
    BH (>= 1.66.0),
    Rcpp (>= 0.12.0),
    RcppEigen (>= 0.3.3.3.0),
    RcppParallel (>= 5.0.1),
    rstan (>= 2.26.23),
    StanHeaders (>= 2.26.22)
SystemRequirements: GNU make
```

This package runs Bayesian statistical models written in Stan using data from paired environmental DNA (eDNA) and traditional surveys provided by a user. Additional functions support interpretation of model output.

Not yet.

The target audience is environmental science researchers or managers who want to interpret environmental DNA data but do not have experience writing and implementing custom Bayesian models.

To my knowledge, no other R package implements similar types of models. Although the R package contains additional model variations, the model was originally developed in a publication: "Tracking an invasion front with environmental DNA", led by the authors of this submission.

Yes
