### 2024-02-22 #628 @ldecicco-USGS
Hi @abigailkeller
I'm new to the editor process for the stats submissions, so sorry for the delay.

The stat submission guide is here:
https://stats-devguide.ropensci.org/pkgdev.html

I think the next step for you is to follow the directions here:
https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-srr
https://github.com/ropensci-review-tools/srr

The eDNAjoint would need srr results to >50% before we can move the package to a full submission.
