### 2024-02-27 #628 @mpadge
Hi @abigailkeller, Mark here from the rOpenSci stats team, to politely disagree with you about your package being out of scope. Scope decisions for stats packages are pretty straightforward: If you think that your package can comply with at least half of all applicable standards - in your case [General](https://stats-devguide.ropensci.org/standards.html#standards-general) and [Bayesian](https://stats-devguide.ropensci.org/standards.html#standards-bayesian), then it is by definition in scope.

Please feel free to have a look through your code, documentation, and tests, check all standards you think will likely apply, and if that seems like more than half, you are very welcome to open this issue again. Then, once you've documented compliance, ask the bot to call `check srr` yourself. As long as that requirement is met, the submission may then proceed. Thanks :smile:
