+++ R/jointModel.R
@@ -410,9 +410,25 @@ init_joint <- function(n.chain,count_all){
 ################
 #helper functions: input checks
 ################
+#' @srrstats {G5.2a} Pre-processing routines to check inputs have unique messages
 
 # if q = TRUE
 catchability_checks <- function(data,cov){
+
+  ## make sure count.type is not zero-length
+  #' @srrstats {G5.8,G5.8a} Pre-processing routines to check for zero-length data
+  if (dim(data$count.type)[1]==0) {
+    errMsg = paste("count.type contains zero-length data.")
+    stop(errMsg)
+  }
+  ## make sure no column is entirely NA in count.type
+  #' @srrstats {G5.8,G5.8c} Pre-processing routines to check for column with all NA
+  if (any(apply(data$count.type, 2, function(col) all(is.na(col))))) {
+    errMsg = paste("count.type contains a column with all NA.")
+    stop(errMsg)
+  }
+
+  #' @srrstats {G2.13} Pre-processing routines to check for missing data
   ## All tags in data are valid (i.e., include qPCR.N, qPCR.K, count, count.type, and site.cov)
   #cov='None'
   if (all(cov=='None') && !all(c('qPCR.N', 'qPCR.K', 'count','count.type') %in% names(data))){
@@ -424,14 +440,16 @@ catchability_checks <- function(data,cov){
     errMsg = paste("Data should include 'qPCR.N', 'qPCR.K', 'count', 'count.type', and 'site.cov'.")
     stop(errMsg)
   }
+
   ## make sure dimensions of count and count.type are equal, if count.type is present
-  #' @srrstats {BS2.1} Pre-processing routines to ensure all input data is dimensionally commensurate
+  #' @srrstats {BS2.1,G2.13} Pre-processing routines to ensure all input data is dimensionally commensurate
   if(dim(data$count)[1]!=dim(data$count.type)[1]|dim(data$count)[2]!=dim(data$count.type)[2]) {
       errMsg = paste("Dimensions of count and count.type do not match.")
       stop(errMsg)
   }
 
   ## make sure all data is numeric -- if q == TRUE
+  #' @srrstats {G5.8,G5.8b} Pre-processing routines to check for data of unsupported type
   if(is.numeric(data$qPCR.K)==FALSE |
        is.numeric(data$qPCR.N)==FALSE |
        is.numeric(data$count)==FALSE |
@@ -440,12 +458,12 @@ catchability_checks <- function(data,cov){
       stop(errMsg)
   }
   ## make sure locations of NAs in count data match locations of NAs in count.type data
-  #' @srrstats {BS2.1} Pre-processing routines to ensure all input data is dimensionally commensurate
+  #' @srrstats {BS2.1,G2.13} Pre-processing routines to ensure all input data is dimensionally commensurate
   if(sum(is.na(data$count.type))!=sum(is.na(data$count))){
       errMsg = paste("Empty data cells (NA) in count data should match empty data cells (NA) in count.type data.")
       stop(errMsg)
   }
-  #' @srrstats {BS2.1} Pre-processing routines to ensure all input data is dimensionally commensurate
+  #' @srrstats {BS2.1,G2.13} Pre-processing routines to ensure all input data is dimensionally commensurate
   if(any((which(is.na(data$count))==which(is.na(data$count.type)))==FALSE)){
       errMsg = paste("Empty data cells (NA) in count data should match empty data cells (NA) in count.type data.")
       stop(errMsg)
@@ -457,6 +475,7 @@ catchability_checks <- function(data,cov){
   }
 
[…]
