+++ R/traditionalModel.R
@@ -231,21 +231,38 @@ init_trad <- function(n.chain, count_all){
 }
 
 # function for input checks
+#' @srrstats {G5.2a} Pre-processing routines to check inputs have unique messages
 traditionalModel_input_checks <- function(data, family, q, phipriors){
-  ## #1. make sure all data tags are valid -- if q == TRUE
+
+  ## make sure all data tags are valid -- if q == TRUE
+  #' @srrstats {G2.13} Pre-processing routines to check for missing data
   if (q==TRUE && !all(c('count.type','count') %in% names(data))){
     errMsg = paste("Data should include 'count' and 'count.type'.")
     stop(errMsg)
   }
 
-  ## #2. make sure all data tags are valid -- if q == FALSE
+  ## make sure all data tags are valid -- if q == FALSE
+  #' @srrstats {G2.13} Pre-processing routines to check for missing data
   if (q==FALSE && !all(c('count') %in% names(data))){
     errMsg = paste("Data should include 'count'.")
     stop(errMsg)
   }
 
-  ## #3. make sure dimensions of count and count.type are equal, if count.type is present
-  #' @srrstats {BS2.1} Pre-processing routines to ensure all input data is dimensionally commensurate
+  ## make sure count is not zero-length
+  #' @srrstats {G5.8,G5.8a} Pre-processing routines to check for zero-length data
+  if (dim(data$count)[1]==0) {
+    errMsg = paste("count contains zero-length data.")
+    stop(errMsg)
+  }
+  ## make sure no column is entirely NA in count
+  #' @srrstats {G5.8,G5.8c} Pre-processing routines to check for column with all NA
+  if (any(apply(data$count, 2, function(col) all(is.na(col))))) {
+    errMsg = paste("count contains a column with all NA.")
+    stop(errMsg)
+  }
+
+  ## make sure dimensions of count and count.type are equal, if count.type is present
+  #' @srrstats {BS2.1,G2.13} Pre-processing routines to ensure all input data is dimensionally commensurate
   if (q==TRUE){
     if(dim(data$count)[1]!=dim(data$count.type)[1]|dim(data$count)[2]!=dim(data$count.type)[2]) {
       errMsg = paste("Dimensions of count and count.type do not match.")
@@ -253,8 +270,16 @@ traditionalModel_input_checks <- function(data, family, q, phipriors){
     }
   }
 
-  ## #4. make sure all data is numeric -- if q == TRUE
+  ## make sure no data are undefined
+  #' @srrstats {G2.16} Pre-processing routines to check for undefined data
+  if(any(data$count==Inf) | any(data$count==-Inf)){
+    errMsg = paste("count contains undefined values (i.e., Inf or -Inf)")
+    stop(errMsg)
+  }
+
+  ## make sure all data is numeric -- if q == TRUE
   #' @srrstats {BS2.5} Checks of appropriateness of numeric values submitted for distributional parameters (i.e., count data must numeric), implemented prior to analytic routines
+  #' @srrstats {G5.8,G5.8b} Pre-processing routines to check for data of unsupported type
   if (q==TRUE) {
     if(is.numeric(data$count)==FALSE |
        is.numeric(data$count.type)==FALSE) {
@@ -263,8 +288,9 @@ traditionalModel_input_checks <- function(data, family, q, phipriors){
     }
   }
 
-  ## #5. make sure all data is numeric -- if q == FALSE
+  ## make sure all data is numeric -- if q == FALSE
   #' @srrstats {BS2.5} Checks of appropriateness of numeric values submitted for distributional parameters (i.e., count data must positive and numeric), implemented prior to analytic routines
[…]
