+++ R/detectionCalculate.R
@@ -39,10 +39,10 @@
 #'                      family='poisson', p10priors=c(1,20), q=FALSE)
 #'
 #' # Calculate at the mean covariate values (covariates are standardized, so mean=0)
-#' detectionCalculate(fit.cov, mu = seq(from=0.1,to=1,by=0.1), cov.val = c(0,0), qPCR.N = 3)
+#' detectionCalculate(fit.cov$model, mu = seq(from=0.1,to=1,by=0.1), cov.val = c(0,0), qPCR.N = 3)
 #'
 #' # Calculate mu_critical at habitat size 0.5 z-scores greater than the mean
-#' detectionCalculate(fit.cov, mu = seq(from=0.1,to=1,by=0.1), cov.val = c(0,0.5), qPCR.N = 3)
+#' detectionCalculate(fit.cov$model, mu = seq(from=0.1,to=1,by=0.1), cov.val = c(0,0.5), qPCR.N = 3)
 #'
 #' # Ex. 2: Calculating necessary effort for detection with multiple traditional gear types
 #'
@@ -54,10 +54,10 @@
 #'                    p10priors=c(1,20), q=TRUE)
 #'
 #' # Calculate
-#' detectionCalculate(fit.q, mu = seq(from=0.1,to=1,by=0.1), cov.val = 'None', qPCR.N = 3)
+#' detectionCalculate(fit.q$model, mu = seq(from=0.1,to=1,by=0.1), cov.val = 'None', qPCR.N = 3)
 #'
 #' # Change probability of detecting presence to 0.95
-#' detectionCalculate(fit.q, mu = 0.1, cov.val = 'None', probability = 0.95, qPCR.N = 3)
+#' detectionCalculate(fit.q$model, mu = 0.1, cov.val = 'None', probability = 0.95, qPCR.N = 3)
 #' }
 #'
 
@@ -663,45 +663,54 @@ detectionCalculate_input_checks <- function(modelfit, mu, cov.val, probability){
   ## #1. make sure model fit is of class stanfit
   #' @srrstats {G2.8} Makes sure input of sub-function is of class 'stanfit' (i.e., output of jointModel())
   if(!is(modelfit,'stanfit')) {
-    errMsg = paste("modelfit must be of class 'stanfit'.")
+    errMsg = "modelfit must be of class 'stanfit'."
     stop(errMsg)
   }
 
   ## #2. make sure mu is a numeric vector
+  #' @srrstats {G5.8,G5.8b} Pre-processing routines to check for data of unsupported type
   if(!any(is.numeric(mu)) | any(mu <= 0)) {
-    errMsg = paste("mu must be a numeric vector of positive values")
+    errMsg = "mu must be a numeric vector of positive values"
     stop(errMsg)
   }
 
   ## #3. make sure probability is a numeric value between 0 and 1
   #' @srrstats {G2.0,G2.2} Assertion on length of input, prohibit multivariate input to parameters expected to be univariate
   if(!is.numeric(probability) | length(probability)>1 | any(probability < 0) | any(probability > 1)) {
-    errMsg = paste("probability must be a numeric value between 0 and 1")
+    errMsg = "probability must be a numeric value between 0 and 1"
     stop(errMsg)
   }
 
   ## #4. cov.val is numeric, if provided
+  #' @srrstats {G5.8,G5.8b} Pre-processing routines to check for data of unsupported type
   if(all(cov.val != 'None') && !is.numeric(cov.val)) {
-    errMsg = paste("cov.val must be a numeric vector")
+    errMsg = "cov.val must be a numeric vector"
     stop(errMsg)
   }
 
   ## #5. Only include input cov.val if covariates are included in model
   if(all(cov.val != 'None') && !c('alpha') %in% modelfit@model_pars) {
[…]
