+++ R/traditionalModel.R
@@ -4,20 +4,28 @@
 #'
 #' @srrstats {G1.4} Roxygen function documentation begins here
 #' @export
-#' @srrstats {BS1.1,BS3.0} Descriptions of how to enter data, description of how NAs are handled
+#' @srrstats {BS1.1,BS3.0,G2.14} Descriptions of how to enter data, description of how NAs are handled (which are informative and should be deliberately included in input data, where necessary)
 #' @param data A list containing data necessary for model fitting. Valid tags are `count` and `count.type`. `count` is a matrix or data frame of traditional survey count data, with first dimension equal to the number of sites (i) and second dimension equal to the maximum number of traditional survey replicates at a given site (j). `count.type` is an optional matrix or data frame of integers indicating gear type (k) used in corresponding count data, with first dimension equal to the number of sites (i) and second dimension equal to the maximum number of traditional survey replicates at a given site (j). Values should be integers beginning with 1 (referring to the first gear type) to n (last gear type). Empty cells should be NA and will be removed during processing. Sites, i, should be consistent in all count data.
 #' @param family The distribution class used to model traditional survey count data. Options include poisson ('poisson'), negative binomial ('negbin'), and gamma ('gamma'). Default value is 'poisson'.
 #' @param q A logical value indicating whether to estimate a catchability coefficient, q, for traditional survey gear types (TRUE) or to not estimate a catchability coefficient, q, for traditional survey gear types (FALSE). Default value is FALSE.
 #' @srrstats {BS1.0,G2.1a,BS1.2} Description of hyperparameters and how to specify prior distributions, explicit documentation of vector input types
 #' @param phipriors A numeric vector indicating gamma distribution hyperparameters (shape, rate) used as the prior distribution for phi, the overdispersion in the negative binomial distribution for traditional survey gear data. Used when family = 'negbin.' Default vector is c(0.25,0.25).
 #' @param multicore A logical value indicating whether to parallelize chains with multiple cores. Default is TRUE.
+#' @srrstats {BS2.7,BS2.11} Option for user to provide initial values for each chain
+#' @param initial_values A list of lists of initial values to use in MCMC. The length should equal the number of MCMC chains. Initial values can be provided for parameters: mu and q. If no initial values are provided, default random values are drawn.
 #' @srrstats {BS1.3} Description of parameters used in the computational process begins here
 #' @param n.chain Number of MCMC chains. Default value is 4.
 #' @param n.iter.burn Number of warm-up MCMC iterations. Default value is 500.
 #' @param n.iter.sample Number of sampling MCMC iterations. Default value is 2500.
 #' @param thin A positive integer specifying the period for saving samples. Default value is 1.
 #' @param adapt_delta Target average acceptance probability used in `rstan::sampling`. Default value is 0.9.
-#' @return An object of class `stanfit` returned by `rstan::sampling`
+#' @srrstats {BS5.0} function returns seeds and initial values used in computation
+#' @return A list of:
+#' \itemize{
+#' \item a model object of class `stanfit` returned by `rstan::sampling`
+#' \item seeds used in MCMC chains
+#' \item initial values used in MCMC
+#' }
 #'
 #' @note  Before fitting the model, this function checks to ensure that the model specification is possible given the data files. These checks include:
 #' \itemize{
@@ -36,7 +44,6 @@
 #' @examples
 #' \donttest{
 #' # Load data
-#' @srrstats {BS1.1} Descriptions of how to enter data
[…]
