+++ R/jointModel.R
@@ -24,11 +24,10 @@
 #' @param n.iter.sample Number of sampling MCMC iterations. Default value is 2500.
 #' @param thin A positive integer specifying the period for saving samples. Default value is 1.
 #' @param adapt_delta Target average acceptance probability used in `rstan::sampling`. Default value is 0.9.
-#' @srrstats {BS5.0} function returns seeds and initial values used in computation
+#' @srrstats {BS5.0} function returns initial values used in computation
 #' @return A list of:
 #' \itemize{
 #' \item a model object of class `stanfit` returned by `rstan::sampling`
-#' \item seeds used in MCMC chains
 #' \item initial values used in MCMC
 #' }
 #'
@@ -119,7 +118,7 @@ jointModel <- function(data, cov='None', family='poisson', p10priors=c(1,20), q=
   all_checks(data,cov,family,p10priors,phipriors)
 
   # initial value checks
-  if(initial_values != 'None'){
+  if(all(initial_values != 'None')){
     initial_values_checks(initial_values,data,cov,n.chain)
   }
 
@@ -183,7 +182,7 @@ jointModel <- function(data, cov='None', family='poisson', p10priors=c(1,20), q=
   #if present, prepare covariate data
   if(all(cov!='None')){
     #' @srrstats {G2.7,G2.10} Use as.data.frame() to allow input list of any tabular form (i.e., matrix, etc.) and converts before filtering columns based on input 'cov'
-    site_mat <- as.data.frame(data$site.cov)[,cov]
+    site_mat <- as.data.frame(as.data.frame(data$site.cov)[,cov])
     site_mat <- cbind(as.data.frame(rep(1,length(site_mat[,1]))),site_mat)
     colnames(site_mat) <- c('int',cov)
   }
@@ -221,14 +220,6 @@ jointModel <- function(data, cov='None', family='poisson', p10priors=c(1,20), q=
     cores <- 1
   }
 
-  # prepare seeds
-  seeds <- sample(1:1000000,n.chain)
-  #' @srrstats {BS2.9} Ensure each chain is started with a different seed by default
-  # ensure seeds are not equal
-  while (any(duplicated(seeds))) {
-    seeds <- sample(1:1000000,n.chain)
-  }
-
 
   ##run model, catchability, pois/gamma, no covariates
   if(q==TRUE&&family!='negbin'&&all(cov=='None')){
@@ -245,7 +236,6 @@ jointModel <- function(data, cov='None', family='poisson', p10priors=c(1,20), q=
                            cores = cores,
                            #' @srrstats {G2.4,G2.4a} explicit conversion to integers for sampling arguments
                            chains = as.integer(n.chain),
-                           seed = seeds,
                            thin = as.integer(thin),
                            warmup = as.integer(n.iter.burn),
                            iter = as.integer(n.iter.burn) + as.integer(n.iter.sample),
@@ -264,7 +254,6 @@ jointModel <- function(data, cov='None', family='poisson', p10priors=c(1,20), q=
                            cores = cores,
                            #' @srrstats {G2.4,G2.4a} explicit conversion to integers for sampling arguments
                            chains = as.integer(n.chain),
-                           seed = seeds,
                            thin = as.integer(thin),
                            warmup = as.integer(n.iter.burn),
                            iter = as.integer(n.iter.burn) + as.integer(n.iter.sample),
@@ -281,7 +270,6 @@ jointModel <- function(data, cov='None', family='poisson', p10priors=c(1,20), q=
                            cores = cores,
[…]
