+++ R/detectionCalculate.R
@@ -65,7 +65,7 @@ detectionCalculate <- function(modelfit, mu, cov.val = 'None', probability=0.9, 
 
   # input checks
   #' @srrstats {G2.1} Types of inputs are checked/asserted using this helper function
-  detectionCalculate_input_checks(modelfit, mu, cov.val, probability)
+  detectionCalculate_input_checks(modelfit, mu, cov.val, probability, qPCR.N)
 
   if (!requireNamespace("rstan", quietly = TRUE)){
     stop ("The 'rstan' package is not installed.", call. = FALSE)
@@ -659,7 +659,7 @@ div_check <- function(x){
 
 # function for input checks
 #' @srrstats {G5.2a} Pre-processing routines to check inputs have unique messages
-detectionCalculate_input_checks <- function(modelfit, mu, cov.val, probability){
+detectionCalculate_input_checks <- function(modelfit, mu, cov.val, probability, qPCR.N){
   ## #1. make sure model fit is of class stanfit
   #' @srrstats {G2.8} Makes sure input of sub-function is of class 'stanfit' (i.e., output of jointModel())
   if(!is(modelfit,'stanfit')) {
