+++ man/jointModel.Rd
@@ -17,7 +17,8 @@ jointModel(
   n.iter.burn = 500,
   n.iter.sample = 2500,
   thin = 1,
-  adapt_delta = 0.9
+  adapt_delta = 0.9,
+  verbose = TRUE
 )
 }
 \arguments{
@@ -46,6 +47,8 @@ jointModel(
 \item{thin}{A positive integer specifying the period for saving samples. Default value is 1.}
 
 \item{adapt_delta}{Target average acceptance probability used in \code{rstan::sampling}. Default value is 0.9.}
+
+\item{verbose}{Logical value controlling the verbosity of output (i.e., warnings, messages, progress bar). Default is TRUE.}
 }
 \value{
 A list of:

[…]
