### 2024-03-11 #628 @abigailkeller
Re-opened issue. I believe my package is now passing at least 55% of the general statistics and Bayesian statistics standards. Thanks!
