+++ DESCRIPTION
@@ -4,7 +4,7 @@ Version: 0.1
 Maintainer: Abigail G. Keller <agkeller@berkeley.edu>
 Authors@R: 
     c(person("Abigail G.", "Keller", role = c("aut", "cre"), email="agkeller@berkeley.edu"),
-    person("Ryan P.", "Kelly", role = "aut", email="rpkelly@uw.edu"))
+    person("Ryan P.", "Kelly", role = "ctb", email="rpkelly@uw.edu"))
 Description: Models integrate environmental DNA (eDNA) detection data and traditional survey data to jointly estimate species catch rate. Models can be used with count data via traditional survey methods (i.e., trapping, electrofishing, visual) and replicated eDNA detection/nondetection data via polymerase chain reaction (i.e., PCR or qPCR) from multiple survey locations. Estimated parameters include probability of a false positive eDNA detection, a site-level covariates that scale the sensitivity of eDNA surveys relative to traditional surveys, and catchability coefficients for traditional gear types. Models are implemented with a Bayesian framework (Markov chain Monte Carlo) using the 'Stan' probabilistic programming language.
 License: GPL-3
 Encoding: UTF-8
