### 2024-03-12 #628 @abigailkeller
Just to add to the ropensci-review-bot:
I believe the missing standards (G1.5, G5.4, G5.5,BS1.2b, BS1.2c, BS1.4, BS4.0, BS4.3, BS6.2, BS6.3) are in the .Rmd file in the vignette folder
