### 2024-03-13 #628 @mpadge
@ldecicco-USGS That `srr` check by the bot should now be correct. @abigailkeller You can call [`srr::srr_stats_pre_submit()`](https://docs.ropensci.org/srr/reference/srr_stats_pre_submit.html) locally to confirm that your package will pass, and then ask the bot here to `check srr` when you think you've finished documenting. Sorry both for any inconvenience here.
