### 2024-03-14 29cd66d
working on more bayesian standards tests
