+++ tests/testthat/test-jointModel.R
@@ -407,11 +407,16 @@ test_that("jointModel input checks work", {
 })
 
 
-# parameter recovery tests
-#' @srrstats {G5.6} Parameter recovery tests to test that the implementation produces expected results given data with known properties
+# correctness and parameter recovery tests
+#' @srrstats {G5.4, G5.6} Correctness/parameter recovery tests to test that the implementation produces expected results given data with known properties
 test_that("jointModel parameter recovery tests work",{
 
+  ################################
+  # model run 1: smaller dataset #
+  ################################
+
   # simulate data: seed 123
+  #' @srrstats {G5.5, G5.6} Running correctness test with a fixed random seed
   set.seed(123)
   # constants
   nsite <- 20
@@ -493,6 +498,9 @@ test_that("jointModel parameter recovery tests work",{
   # all should be equal to true
   expect_equal(check,rep(TRUE,nsite+length(alpha)+length(log_p10)))
 
+  ##########################################
+  # model run 2: smaller dataset, new seed #
+  ##########################################
 
   #' @srrstats {G5.6b,G5.9b} Run test for multiple seeds with same data
   # simulate data: seed 222
@@ -552,6 +560,92 @@ test_that("jointModel parameter recovery tests work",{
   for(i in 1:nsite){
     par <- paste0('mu[',i,']')
     if(mu[i] > summary[par,'2.5%'] && mu[i] < summary[par,'97.5%']){
+      check[i] <- TRUE
+    } else {
+      check[i] <- FALSE
+    }
+  }
+  # check alpha
+  for(i in 1:length(alpha)){
+    par <- paste0('alpha[',i,']')
+    if(alpha[i] > summary[par,'2.5%'] && alpha[i] < summary[par,'97.5%']){
+      check[nsite+i] <- TRUE
+    } else {
+      check[nsite+i] <- FALSE
+    }
+  }
+  # check p10
+  if(log_p10 > summary['log_p10','2.5%'] && log_p10 < summary['log_p10','97.5%']){
+    check[nsite+length(alpha)+1] <- TRUE
+  } else {
+    check[nsite+length(alpha)+1] <- FALSE
+  }
+
+  #' @srrstats {G3.0, G5.6a} Instead of comparing floating point values for equality, here the model is tested to determine if the true parameter values are within the 95% quantiles of the posterior
+  # all should be equal to true
+  expect_equal(check,rep(TRUE,nsite+length(alpha)+length(log_p10)))
+
+  ################################
+  # model run 3: larger dataset #
+  ################################
+
+  # simulate data: seed 123
+  set.seed(123)
+  # constants
+  nsite <- 40 # increased dataset size
+  nobs_count <- 400 # increased dataset size
+  nobs_pcr <- 20 # increased dataset size
+  # params
+  mu <- rlnorm(nsite,meanlog=log(1),sdlog=1)
+  alpha <- c(0.5, 0.1, -0.4)
+  log_p10 <- -4.5
+  # count
+  count <- matrix(NA,nrow=nsite,ncol=nobs_count)
+  for(i in 1:nsite){
+    count[i,] <- rpois(nobs_count,mu[i])
+  }
+  # site-level covariates
+  mat_site <- matrix(NA,nrow=nsite,ncol=length(alpha))
+  mat_site[,1] <- 1 # intercept
+  for(i in 2:length(alpha)){
+    mat_site[,i] <- rnorm(nsite,0,1)
+  }
[…]
