### 2024-03-19 #628 @ldecicco-USGS
Thank you @abigailkeller for this pre-submission inquiry! This package is well within our statistical scope and we'd be happy to review a full submission. So now you'll need to open a new issue for the full submission. Let me know if you have any questions.

(see issue 615 for a recent example of a submission, where the pre-submission is 614)
